requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity
azure-storage-blob
modal_core_python==0.9.3
pypomes_http==0.3.8
pypomes_logging==0.3.2

